barba transitions

+++ .stylelintrc.js
@@ -3,7 +3,7 @@ module.exports = {
   'defaultSeverity': 'warning',
   'rules': {
     'no-empty-source': null,
-    'string-quotes': 'double',
+    'string-quotes': 'single',
     'at-rule-no-unknown': [
       true,
       {
